requirements.txt
psycopg2-binary>=2.9.9
pandas>=2.2.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
tenacity>=8.2.3
streamlit>=1.40.0
